feat: provide fluent builder so that logforth can be used even for simple demo scenarios
